fix(teams): harden router after security/perf audit
- Remove public webhook-sink endpoint (leaked all request headers + bodies)
- Plug isPrivate member-list leak: members no longer in SSR payload for
  /team/[slug], fetched separately only when not private
- Cap logoUrl/bannerUrl in update schema (1 MB) + sensible caps on bio,
  brandColor, theme; rename memberId → userId in remove/changeMemberRole
- Silence inviteMember (uniform {ok:true} reply, no enumeration oracle, no
  raw token in response, $transaction around token+membership creation)
- Replace N+1 in getActiveUserBreakdown with single groupBy; replace serial
  upserts in addMembersToEventTypes with createMany({skipDuplicates:true});
  push isOrganization filter into Prisma where for list.handler
- Tighten removeMember: only OWNER may kick another ADMIN/OWNER, with last-
  owner guard; delete + host detach in $transaction
- Catch P2002 in create/update.handler for slug uniqueness (race-tolerant)
- Add server page.tsx wrappers (session check) for all teams settings
  pages; client logic moved to view.tsx — closes architecture-page-level-auth
- Profile view: hydrated ref guards initial form hydrate so refetch() after
  save no longer clobbers in-progress edits
- Members view: load-more pagination (50/chunk), confirm before role change
- EventTeamAssignmentTab: extract HostRow as React.memo + stable callbacks
  so a single keystroke only re-renders the affected row
- webhook/util.ts: explicit FORBIDDEN on owner-less orphan webhooks
- /team/[slug]/[type] SSR aligned on isOrganization filter (was parentId)
- Drop dead flags team-booking-page-cache + use-api-v2-for-team-slots
- permissions.ts: accept optional user arg for upcoming admin bypass
