+++ apps/web/modules/feature-flags/hooks/useFlags.ts
@@ -24,8 +24,6 @@ const initialData: AppFlags = {
   "hwm-seating": false,
   "signup-watchlist-review": false,
   "sink-shortener": false,
-  "team-booking-page-cache": false,
-  "use-api-v2-for-team-slots": false,
 };
 
 export function useFlags(): Partial<AppFlags> {
