+++ packages/trpc/server/routers/viewer/teams/changeMemberRole.handler.ts
@@ -8,19 +8,18 @@ import type { TrpcSessionUser } from "../../../types";
 
 type Options = {
   ctx: { user: NonNullable<TrpcSessionUser> };
-  input: { teamId: number; memberId: number; role: MembershipRole };
+  input: { teamId: number; userId: number; role: MembershipRole };
 };
 
 export async function changeMemberRoleHandler({ ctx, input }: Options) {
-  await requireMember(ctx.user.id, input.teamId, MembershipRole.OWNER);
+  await requireMember(ctx.user.id, input.teamId, MembershipRole.OWNER, ctx.user);
 
-  const target = await getMembership(input.memberId, input.teamId);
+  const target = await getMembership(input.userId, input.teamId);
   if (!target) throw new TRPCError({ code: "NOT_FOUND" });
 
-  // Demoting an OWNER is fine, but never end up with zero owners.
   if (target.role === MembershipRole.OWNER && input.role !== MembershipRole.OWNER) {
     const otherOwners = await prisma.membership.count({
-      where: { teamId: input.teamId, role: MembershipRole.OWNER, NOT: { userId: input.memberId } },
+      where: { teamId: input.teamId, role: MembershipRole.OWNER, NOT: { userId: input.userId } },
     });
     if (otherOwners === 0) {
       throw new TRPCError({ code: "BAD_REQUEST", message: "Cannot demote the last owner" });
@@ -28,7 +27,7 @@ export async function changeMemberRoleHandler({ ctx, input }: Options) {
   }
 
   return prisma.membership.update({
-    where: { userId_teamId: { userId: input.memberId, teamId: input.teamId } },
+    where: { userId_teamId: { userId: input.userId, teamId: input.teamId } },
     data: { role: input.role },
     select: { id: true, role: true },
   });
