+++ packages/trpc/server/routers/viewer/teams/getActiveUserBreakdown.handler.ts
@@ -14,7 +14,7 @@ type Options = {
  * Used for the "active per member" chart on the team dashboard.
  */
 export async function getActiveUserBreakdownHandler({ ctx, input }: Options) {
-  await requireMember(ctx.user.id, input.teamId);
+  await requireMember(ctx.user.id, input.teamId, undefined, ctx.user);
 
   const members = await prisma.membership.findMany({
     where: { teamId: input.teamId, accepted: true },
@@ -23,20 +23,24 @@ export async function getActiveUserBreakdownHandler({ ctx, input }: Options) {
     },
   });
 
-  const counts = await Promise.all(
-    members.map(async ({ user }) => {
-      const bookings = await prisma.booking.count({
-        where: {
-          userId: user.id,
-          eventType: { teamId: input.teamId },
-          startTime: { gte: new Date() },
-          status: BookingStatus.ACCEPTED,
-        },
-      });
-      return { user, bookings };
-    })
-  );
+  // Single aggregate query instead of one COUNT per member.
+  const grouped = await prisma.booking.groupBy({
+    by: ["userId"],
+    where: {
+      userId: { in: members.map((m) => m.user.id) },
+      eventType: { teamId: input.teamId },
+      startTime: { gte: new Date() },
+      status: BookingStatus.ACCEPTED,
+    },
+    _count: { _all: true },
+  });
 
-  counts.sort((a, b) => b.bookings - a.bookings);
-  return counts;
+  const countsByUserId = new Map<number, number>();
+  for (const row of grouped) {
+    if (row.userId !== null) countsByUserId.set(row.userId, row._count._all);
+  }
+
+  return members
+    .map(({ user }) => ({ user, bookings: countsByUserId.get(user.id) ?? 0 }))
+    .sort((a, b) => b.bookings - a.bookings);
 }
