+++ packages/trpc/server/routers/viewer/teams/schemas.ts
@@ -9,28 +9,38 @@ export const ZCreateInput = z.object({
   slug: z.string().min(1).max(64),
 });
 
+// Caps a `logoUrl` / `bannerUrl` payload. Web URLs are far under 2 KB, base64
+// data URLs from the ImageUploader (512×512 PNG) land around 250–500 KB.
+// 1 MB leaves headroom for a slightly larger crop without letting a team
+// admin DoS the row by stuffing megabytes of base64 into the column.
+const MAX_IMAGE_FIELD_LENGTH = 1_048_576;
+
+const imageField = z.string().max(MAX_IMAGE_FIELD_LENGTH).nullable().optional();
+
 export const ZUpdateInput = z.object({
   teamId: z.number().int().positive(),
   name: z.string().min(1).max(80).optional(),
   slug: z.string().min(1).max(64).optional(),
-  bio: z.string().nullable().optional(),
-  logoUrl: z.string().nullable().optional(),
-  bannerUrl: z.string().nullable().optional(),
-  brandColor: z.string().nullable().optional(),
-  darkBrandColor: z.string().nullable().optional(),
-  theme: z.string().nullable().optional(),
+  bio: z.string().max(8_000).nullable().optional(),
+  logoUrl: imageField,
+  bannerUrl: imageField,
+  brandColor: z.string().max(32).nullable().optional(),
+  darkBrandColor: z.string().max(32).nullable().optional(),
+  theme: z.string().max(64).nullable().optional(),
   isPrivate: z.boolean().optional(),
   hideBranding: z.boolean().optional(),
 });
 
 export const ZRemoveMemberInput = z.object({
   teamId: z.number().int().positive(),
-  memberId: z.number().int().positive(),
+  // Despite the historic `memberId` name, this is the *user* id (the
+  // Membership compound key is (userId, teamId), not its own row id).
+  userId: z.number().int().positive(),
 });
 
 export const ZChangeMemberRoleInput = z.object({
   teamId: z.number().int().positive(),
-  memberId: z.number().int().positive(),
+  userId: z.number().int().positive(),
   role: z.nativeEnum(MembershipRole),
 });
 
@@ -43,6 +53,10 @@ export const ZInviteMemberInput = z.object({
 export const ZAcceptOrLeaveInput = z.object({
   teamId: z.number().int().positive(),
   accept: z.boolean(),
+});
+
+export const ZAdminListInput = z.object({
+  search: z.string().max(128).optional(),
 });
 
 export const ZAddMembersToEventTypesInput = z.object({
